Move MiningOptions to upper package
